tools/rocs-cli/pyproject.toml
[project]
name = "rocs-cli"
version = "0.1.4"
description = "Minimal ROCS CLI for ai-society"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.0",
  "rich>=13.0.0",
]

[project.scripts]
rocs = "rocs_cli.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
